refactor: anti hyst wl change (#25)
* refactor: sweep mode change strategies and anti-hyst function

The tune mode wavelength changing method was implemented using a
strategy pattern. This was refactored now to just use two functions
improving readability of the code. The anti-hyst procedure was declared
in 2 places. Tune mode and phase section. Now it's just in the phase
section with the tune mode calling it and a constructor func was made to
allow for the method to change upon calibration

* docs: update diagram